[feat/#240] 🎨 Style: 레이아웃 스타일 수정
