+++ chatty-fe/src/app/(game)/layout.css.ts
@@ -4,5 +4,6 @@ export const mainContainer = style({
   padding: '50px 20px',
   display: 'flex',
   width: '100%',
+  height: 'calc(100% - 60px)',
   justifyContent: 'center',
 });
