+++ chatty-fe/src/app/(game)/main-lobby/page.css.ts
@@ -5,8 +5,8 @@ export const lobbyContainer = style({
   flexDirection: 'column',
   width: '100%',
   height: '100%',
-  paddingBottom: 100,
   alignItems: 'center',
+  overflowY: 'hidden',
 });
 
 export const centerContainer = style({
